+++ .vscode/settings.json
@@ -62,6 +62,7 @@
     "govulncheck",
     "graffico",
     "gverrt",
+    "hanno",
     "hiber",
     "icase",
     "ineffassign",
