+++ internal/feat/resume/save_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/snivilised/agenor/internal/services"
 	"github.com/snivilised/agenor/locale"
 	"github.com/snivilised/agenor/pref"
-	"github.com/snivilised/agenor/test/hydra"
+	"github.com/snivilised/agenor/test/hanno"
 	"github.com/snivilised/agenor/tfs"
 	"github.com/snivilised/li18ngo"
 	nef "github.com/snivilised/nefilim"
@@ -93,7 +93,7 @@ var _ = Describe("Save", Ordered, func() {
 
 	BeforeEach(func() {
 		services.Reset()
-		fS = hydra.Nuxx(verbose, lab.Static.RetroWave)
+		fS = hanno.Nuxx(verbose, lab.Static.RetroWave)
 		rS = tfs.New()
 	})
