+++ internal/filtering/filter-regex_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/snivilised/agenor/internal/services"
 	"github.com/snivilised/agenor/internal/third/lo"
 	"github.com/snivilised/agenor/pref"
-	"github.com/snivilised/agenor/test/hydra"
+	"github.com/snivilised/agenor/test/hanno"
 	"github.com/snivilised/agenor/tfs"
 	"github.com/snivilised/li18ngo"
 	"github.com/snivilised/nefilim/test/luna"
@@ -30,7 +30,7 @@ var _ = Describe("feature", Ordered, func() {
 			verbose = false
 		)
 
-		fS = hydra.Nuxx(verbose,
+		fS = hanno.Nuxx(verbose,
 			lab.Static.RetroWave, "PROGRESSIVE-HOUSE",
 		)
 		Expect(li18ngo.Use()).To(Succeed())
