+++ test/hanno/virtual-tree.go
@@ -1,4 +1,4 @@
-package hydra
+package hanno
 
 import (
 	"bytes"
